requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
flake8==6.0.0
flake8-isort==6.0.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
protobuf==4.22.1
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
python-dotenv==1.0.0
requests==2.28.2
rsa==4.9
schedule==1.1.0
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
